chore: Communication classes should be separated
