+++ include/roboutils/Log.h
@@ -9,7 +9,7 @@
 
 #include <iostream>
 #include <string>
-#include "roboutils/UDP.h"
+#include "roboutils/comm/UDP.h"
 
 namespace RoboUtils {
     using namespace std;
